refactor: move to monorepo
