+++ cli/README.md
@@ -1,0 +1,1 @@
+## Wharf CLI
