+++ cli/go.mod
@@ -1,0 +1,9 @@
+module github.com/Makepad-fr/wharf/cli
+
+go 1.22.2
+
+replace github.com/Makepad-fr/wharf/core => ../core
+
+require github.com/Makepad-fr/wharf/core v0.0.0-00010101000000-000000000000
+
+require gopkg.in/yaml.v3 v3.0.1 // indirect
