+++ cli/wharf.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/Makepad-fr/wharf/core"
+)
+
+func main() {
+	fmt.Println(core.Version)
+}
