+++ core/version.go
@@ -1,0 +1,3 @@
+package core
+
+const Version = "0.0.1"
