+++ docker-plugins/commons/plugin_metadata.go
@@ -1,4 +1,4 @@
-package docker_render
+package commons
 
 import (
 	"encoding/json"
@@ -14,16 +14,8 @@ type PluginMetadata struct {
 	Experimental     bool   `json:"Experimental"`
 }
 
-func ShowPluginMetaData() bool {
+func ShowPluginMetaData(metadata PluginMetadata) bool {
 	if len(os.Args) > 1 && os.Args[1] == "docker-cli-plugin-metadata" {
-		metadata := PluginMetadata{
-			SchemaVersion:    "0.1.0",
-			Vendor:           "MAKEPAD",
-			Version:          version,
-			ShortDescription: "A Helmlike templating engine for Docker",
-			URL:              "https://github.com/Makepad-fr/docker-render",
-			Experimental:     true,
-		}
 		json.NewEncoder(os.Stdout).Encode(metadata)
 		return true
 	}
