--- version.go
@@ -1,3 +1,0 @@
-package docker_render
-
-const version = "0.0.1"
